Add files via upload
Task 1.
For the given ranges for OX and OY, create an empty graph.
Task 2.
An example of a quadratic function graph
Task 3.
Set a parametric curve, i.e., the coordinates are given by a pair.
(x(t), y(t)), where tmin ⩽ t ⩽ tmax.
The chart contains a grid, an axis description, and a legend. The axes have the same scale.
Task 4.
Create an arbitrary-different graph of at least 3
Functions, colors, and line types are to be specified.
Task 5.
Let there be a numerical sequence of the form
an = α + β an−1, a0 ≡ const.
For the given parameters and the starting value of the sequence, draw a sequence of N sequence values. Example for a0 = 10
α = 2, β = 0.9: